AndroidX Webkit: reenable flaky test
This reenables a flaky test. Flakiness should have been resolved prior
to 74.0.3726.0. AndroidX pulls in version 74.0.3729.157, so this should
certainly be fixed.

Flakiness was introduced in a chromium update (the test used to not be
flaky), so the bug is not in the system image version of WebView for any
of the non-updating bots.

Fixes: 111690396
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I061e2c96f1ce1131ca1d6a3a816d3baffbef2d8a