Animate Toolbar Changes in NavigationUI
When animateLayoutChanges is set on a Toolbar, there are errors in how
the text within the toolbar is displayed if it has an arrow. The text
leaves a big space on destinations with no arrow.

This change checks so see if the icon will be null and then uses a
TransitionManager to animate the toolbar text into the proper position.
We only animate when the icon is going to disappear. If we were animate
when the icon is appearing, the new icon and text appear stacked when
the animation begins.

Test: Tested visually in app
BUG: 131403621
Change-Id: I97ff213bdeb6da3282287d9b572f33c74cf9e9e7