Allow hosting biometric prompt demo in activity or fragment
Currently, the androidx support biometric prompt demo always launches
the prompt from an activity and hooks into the activity's lifecycle
methods. However, it would be also be useful for testing and
demonstration purposes to launch the prompt from within a fragment as
well.

This commit changes the main screen of the biometric prompt demo app
to prompt the user whether they would like to host the prompt in an
activity or fragment. They can then tap on one of two buttons which then
launches the prompt demo within the appropriate host.

Test: Launch demo app, verify that the first screen presents options to
host the prompt in an activity or fragment. Tap on each of the options,
verify that the demo launches within the appropriate host (as shown in
the label at the top of the screen). Verify all authentication flows
work as before from both the activity and fragment.

Bug: 131980596

Change-Id: I19d520b38a07eba801927d8ce5c12141c40e5839