Add different fit options to Stack
This adds support for specifying how the constraints that the
stack inherits from its parent are passed to its non-positioned
children. The constraints can be loosened, passed unchanged or
tightened to the biggest size allowed. The first two options are
implemented by adding a loose property to aligned children and the
last option by introducing a new type of children called expanded.

Bug: 135670598
Test: Ran StackTest, ConstraintsTest and StackUsage in ComplexLayoutDemos
Change-Id: I6eeb7c421873df6018c1f9b4435049f06de6165a