Fix maxDepVersions testing
Max dep versions was not working (not substituting and went undetected
for a while.) This fixes it. It also removes the need to specify versions
of annotation processors in build.gradle files based on whether we're
running maxdepversions or not. My guess is that's because the
annotationProcessor configuration now has variants (debug/release),
which is probably due to an AGP update, and that made it such that the
substitution understood to replace for debug variants.

Bug: 127495641
Test: gw bOS -PuseMaxDepVersions
Test: gw bOS
Change-Id: I3812189bbcec2d20fa0c045df6c5b58e0dc3e94d