Fall back to commitAllowingStateLoss() on ISE
When FragmentPagerAdapter or FragmentStatePagerAdapter
runs finishUpdate() in response to a measure/layout
pass, it runs commitNowAllowingStateLoss() as all
normal devices don't do measure/layout while processing
pending FragmentTransactions.

However, with b/135945162, FragmentManager does call
requestApplyInsets() as part of processing Fragments.
This gets posted on real devices, but Robolectric will
run it inline, causing this to throw an
IllegalStateException. To ensure that Robolectric
continues to work, we can call commitAllowingStateLoss()
in this particular case to ensure that it runs even
while pending FragmentTransactions are being processed.

Test: manual testing
Fixes: 137201343
Change-Id: I62b6312ffbd754af6c6bedf0f1adc30c40641575