Use ByteBuffer when a relation's key column is of type BLOB.
Relationship resolution via @Relation is done through a Map and when
the key column is of type BLOB a ByteBuffer has to be used, otherwise
using a byte[] would always cause the lookup to fail due to byte[]
using object identity as equals and hashCode.

This CL makes it so that the only during key manipulation and binding
a ByteBuffer is used, which simply wraps the byte[] data of the BLOB
and implements Comparable.

Bug: 137881998
Test: testapp:cC
Change-Id: Iad5d54573b60671467d7e6daa4fbfb52cfa8e8b3