Add API to identify Fragment from view
Currently there is no way to identify if a view is associated with a
Fragment, or what Fragment the view may be associated with.

This changed adds a static API to FragmentManager to allow a Fragment to
be found from a view. The fragment is stored in the view as a tag with
a static id after the fragment view is created, but before the view is
added to the fragment container. The fragment can then be retrieved from
any view that is a child of the fragment's view.

Test: Added Tests in FragmentViewTest, ./gradlew checkApi
BUG: 136494650
Change-Id: Idd79a1bb21c1c62473d1498267ab8ddf5e27cda9