Add Flex cross axis alignment lines positioning
The CL improves Flex to be able to position children in the cross axis
direction according to alignment lines. Currently, if a child does not
provide an alignment line, the Flex will position it at 0.ipx in the
cross axis direction. The API surface is likely to change with
modifiers.

Test: ran FlexTest
Bug: 129859180
Change-Id: Icc29173ee5ab32abc3dbea31b99b8ca5795f694a