Collection[Item]Info should not be populated by default for RecyclerViews.
This is significant number of the incoming bugs I get.
RecyclerView can't accurately determine how many "items" are in the list, or what the "index" of an item is, just because some it's cells may not be exposed to accessibility, or some may not behave like they are a part of the collection.
By default, we shouldn't populate this.
If the app developer chooses to populate it themselves, then we will still respect their decision to do so, and talkback will read that out to the user, just because the app developer is capable of giving us an accurate count and index.
The behavior that will change for talkback users is that, they will no longer here counts for most recyclerViews, unless the app developer has explicity specified counts.

Change-Id: I3023ce09760773c67985e4211d7df13e31e928a9
Fix: 137113167
Test: Support-v7-demos by hand for each of the count conditions (invalid, ambiguous, zero, 1, normal),  RecyclerViewAccessibilityTests, also added unit tests for when Collection[Item]Info is expclity populated, and modified other tests, ViewPager2 sample app, ViewPager 2 accessibility test added some.