Throw IllegalStateException when FragmentManager has no host
When FragmentManager is destroyed or never attached to a host we should
throw IllegalStateExceptions with proper messages to indicate the cause.

This change adds an boolean parameter to execPendingActions() to
indicate whether to allow state loss. This parameter is passed to
ensureExecReady() to decide if state loss needs to be checked.

Test: added new tests
BUG: 64896609
Change-Id: I092218ce5a2930d33f256a651c00c06fb2a24fca