Move ComplexLayoutState to LayoutNode
There's no reason to have two data stores for the same information.
This remove ComplexLayoutState and unifies the logic into
LayoutNode. The average performance for measure/layout is
approximately 20% better based on CheckboxesInRowsTestCase.

This also changes how the varargs Layout works. Instead
of tagging with ParentData and inserting LayoutNodes, it
adds a DataNode for each composable child. This makes
the search for measurables a little faster and composition
just a little more efficient.

Test: ran demo and CheckboxesInRows test
Change-Id: I31a4075b3d367a63c5f72aa83adf88c325b31665