Use ExitAnimationCompleteMarker with Animations and Animators
Animations, Animators, and Transitions are currently all handled
separately by the FragmentManager. Looking toward the future it would be
nice if these and any other animation systems could be handled in the
same (or at least a similar) way by the FragmentManager.

This change uses the ExitAnimationCompleteMarker interface to allow
proper behavior when Animations and Animators are ended or canceled. A
cancel method was added to allow animation components to define how they
should be handled when the FragmentManager needs to cancel an animation
early. When animations are ended, the animation component removes the
ExitAnimationCompleteMarker and that notifies the FragmentManager that
the animation has ended, the view can be destroyed, and the fragment can
be moved to the proper state.

Test: all tests pass
BUG: 138741697
Change-Id: I315b814132a57ab89c0ccd127161b28e60628996