Fix crash on releasing CaptureSession.
Client's test started to crash after Aug 12th release. It happens on real device too when exiting the app. The crash is a NPE from within camera2.

The flakiness was introduced in aosp/980247 where we call abortCaptures() right before calling close(). The theory is that calling abortCaptures() sends ERROR_CAMERA_BUFFER which cleans Surfaces, but it could have already been cleared by close() in a race condition.

The workaround is to call close() in the onReady() callback instead of calling it directly after abortCaptures().

Bug: 139448807
Test: ./gradlew camera:connectedCheck
Change-Id: I9ec57190953a2b2a1ad3bf95cc583ab7b153212d