Override equals() and hashCode() in AccessibilityActionCompat
Make two AccessibilityActionCompat objects equal if they share
the same action ID. AccessibilityAction already does this and it
makes a11y services’ jobs easier when determining if a node
supports an action.

Also get rid of variable mLabel since it’s not being used and
add the rest of the actions to getActionSymbolicName

Bug: b/114119473
Test: Added tests in AccessibilityNodeInfoCompatTest. Tested on R
device.
Change-Id: I4aba64171c18a3a9e2c9700016ed301bacd82f07