Make AnimatedFloat onFinished callback more informative
This CL makes onFinished callback for both DynamicAnimations and
AnimatedFlaot more informative, which unblocks some functionality and
fixes two bugs that occurred only when we use AnimatedFloat as a
multi-shot animation primitive, as opposed to the single shot use case:
-- when the upped bound is reached, animatedFloat end animation with
cancelled true, but in reality there was no cancellation, we just
reached the bound and should stop the animation.
-- when we reach the bound and then user keeps scrolling, we don't reset
the velocity, so even there's no animations happening, velocity keeps
growing and growing, resulting with the wrong/no animation when
animation is happening lately in the opposite direction

Bug: 140169158
Bug: 140169159
Test: all ui-aniumation tests passed, although I can't write proper test
for this bugs because we need some testing infrastructure first
(choreographer, etc).

Change-Id: Ie8d6c9737c5e2d7d615b839c53ed8f4973e96f18