Update CarListDialog to make CarListDialog.Item immutable and remove automatic dismiss on click behavior for undeprecated items.
Automatic dismiss on click was pre-existing behavior that is expected in
existing implementations that use the deprecated setItem() method and
DialogSubSection class. Going forward (when using the undeprecated setItem()
method or new constructor for DialogSubSection), dismissal of the
dialog will require explict dismissal in the click listener or the
inclusion of a dismiss button.

Bug: 140101105
Bug: 140295548
Test: ./gradlew :car:connectedCheck

Change-Id: Ie6e52f8f99e34119310c1ea8bbb38d58984ec329