Add intrinsic measurements to Table
This CL adds two new column width specifications: MinIntrinsic and
MaxIntrinsic. These are similar in behaviour to Wrap, but they use
intrinsic measurements (i.e. minIntrinsicWidth and maxIntrinsicWidth
respectively) and delay measurement until the column widths are known.

This also implements intrinsic measurements for Table itself, which
was necessary to use intrinsic measurements on the children of Table.

Bug: 137850256
Test: Ran TableTest
Test: Added TableTest#testTable_withColumnWidth_[min/max]Intrinsic
Test: Added TableTest#testTable_hasCorrectIntrinsicMeasurements_*
Change-Id: Ib6dc71950cd3d47dff813d19fc96ce983e46ba04