Added assertIsDisplayed function
This function tests if a compoenent is actually displayed on screen
based on it's location in the hierarchy tree, it's width and height,
the screen bounds and if there is clipping enabled.

Currently we do not check if a popup window is displayed on top of the
component nor do we have view compatibility.

Bug: 123702531
Test: androidx.ui.test.IsDisplayedTests

Change-Id: Ic642113cb2f321eefe5a22ccbe789b7c6b1040be