Fix drawing without layout after recomposition
Before this CL, there were cases when a node was asked to draw
themselves without a previous layout. This was happening when an
ancestor was not affecting parent size, and the child was recomposing.
This CL fixes the bug by allowing children to mark parents as
needing layout when subcomposition is happening.
The bug was surfaced after c7a02449443be1ac44b50af6513d32beeaa345a7,
which updated Container to only measure its child in the positioning
block when the size of the Container was known in advance.

Bug: 140292836
Test: ran AndroidLayoutDrawTest
Change-Id: I0c830e0c5dcb84f85f66280b2beafd87127cf02d