[webview] Add renderer termination activity to demo app.
This CL adds an activity to the demo app showcasing renderer termination
APIs. The activity provides the ability to transiently or indefinitely
block the renderer thread. In response (after input events or navigations)
a renderer unresponsive dialog is shown, providing the option to terminate
the renderer or wait. The dialog will be reshown at 5 second intervals
until the renderer is killed, or unblocked.

If the renderer is terminated, a dialog is shown, and the WebView is
recreated in order to recover.

Test: manual; see demo app.
Fixes: 131268134
Change-Id: I90f31dd0d7d1fec58f1ddde3c9eec49fe3e636ae