Allow extra parameters in deep link matching
When matching deep link parameters, parameters not in the original deep
link should be ignored. The parameters we know about should be matched
if they are given, and optional if there is a default value or it is
nullable.

This change removes the check for parameters to be a subset of the
original deep link.

Test: changed test to match new expected behavior
BUG: 141482822
Change-Id: I1a2ec2fbf28bd939b95005200fd2ebfd4211c30c