Only animate toolbar when removing an icon
When animating the toolbar we also animate the bounds. This means if the
toolbar title is long enough, it will transition in. That is okay when
going from a destination with an icon, to destinations with no icon
(top level destination). When going between two top level destinations,
we should avoid animating the toolbar.

Test: Tested visually in dev app, and navigation app
BUG: 140848160
Change-Id: I6b50438bcacbb43f6d9af205a4c88409fcbf6a97