Fix for the shadows ignoring Opacity on Android 10
Due to the bug in Q b/141696413 Android can't automatically detect overlapping drawing when we use RenderNodes so the shadow is drawn as if there will be no opacity applied. To fix it for us we have to set true to hasOverlappingRendering for all our RenderNodes. Maybe we will be able to improve this later if we introduce a mechanism of detecting drawing boundaries in Compose, or when it would be fixed in the framework.

Change-Id: Id22237e31bd3544dcbbf51758b1fc9df88246165
Fixes: 141687011
Test: added a new test for this case