AndroidX Webkit: fix lint issues in WebViewRenderProcessClientAdapter
No change to logic.

This makes WebViewRenderProcessClientAdapter and its methods public so
it may be linted. This fixes nullability lint issues in the internal
class. This removes redundant @hides (the webkit.internal package is
already @hide)

This also fixes false-positive LambdaLast warnings for
setWebViewRenderProcessClient() (in the public API and all the internal
plumbing methods).

Bug: 139770271
Bug: 141775185
Test: ./gradlew :webkit:lintDebug
Change-Id: I4ba238f9ae9e8d511647e0f57154f43766dc6cea