Do not call abortCaptures() when use cases become offline.
This is a follow up on aosp/1131673. Aborting inflight captures improves performance, but it triggers a frameowork NPE on legacy devices. The previous CL disabled it for legacy devices, but it added complexity as well as made it inconsistent on legacy and non-legacy devcies. We have decided to choose consistency & simplicity over performance.

Bug: 139448807
Test: manual test and ./gradlew bOS
Change-Id: If0ac6122a099ff19e686f0f8ffd01ba2e326c97e