Execute DeferrableSurface#getSurface() async
(1) DeferrableSurfaces#getSurface() task to asynchronous method.
(1.1) Add timeout option for DeferrableSurfaces#surfaceList()
(1.2) Add ScheduledExecutorService to schedule the timeout of the
DeferrableSurfaces#surfaceList() method.

(2) Create an extra GET_SURFACE state when executing the
DeferrableSurface#getSurface() task.
(2.1) Add extra ScheduledExecutorService as part of CaptureSession constructor input.

(3) Move the create CaptureSession code to ListenableFuture.

(4) Chaining the task (1) and (3) in CaptureSession#open(),
The listenableFuture of the CaptureSession#open() will be done when the
CaptureSession configured successfully.

Bug: 123723019
Test: ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:camera-extensions:test &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:integration-tests:camera-testapp-extensions:connectedAndroidTest

Change-Id: I94b5c310386ca5aa557db98200b74ae189d2cc25