Ensure interrupting animations start updating value in the current frame
When an animation is interrupted, the value (from last frame) is
being snapshotted and used in the new animation. This means the new
animation should use the last frame time as the start time.
Otherwise, the animation will stutter (i.e. not change value for a
whole frame) until the next frame. Note: This problem was exacerbated
animateTo(sameTargetValue) was called with the same target value on
every frame, as the animation stays forever on the first frame.

Test: calling animateTo() with the same target repeatedly
TODO: once the animationClock is set up, this needs to be a test case

Change-Id: I8c67afacaa8881cc03d0ee83f30ba54a8fd8f7aa