Ensure FragmentContainerView properly inflates Fragments
When given a name, a FragmentContainerView should inflate a Fragment
with the given name and add it to the appropriate FragmentManager. When
the Fragment calls onInflate() the attributes passed from the
FragmentContainerView must be the ones from when the
FragmentContainerView is contructed, otherwise the attributes will be
incorrect.

This change uses a LayoutInflatorFactory to contruct the
FragmentContainerView and pass in the right FragmentManager at
construction time. This means that if a Fragment is created, it can call
onInflate() with the proper AttributesSet.

Test: Added test to mirror all test with the fragment tag. tested in app
BUG: 142421837
Change-Id: I65ff0030a975d3c9f9087660f5b9e4b0a5a6e24c