Fix fragment transition epicenter calculation
The current implementation uses View#getLocationOnScreen, which maps
(0,0) in the view's coordinate system to that of the screen, and then
adds width/height. This produces incorrect results if the view in
question is rotated or scaled (for example, if the view is rotated 180
degrees, the rect would be offset horizontally and vertically by the
view's width/height respectively from where it's actually located).

This implementation takes the view's matrix (and that of its parents)
into account, thus producing a rectangle that accurately corresponds to
the visible position & size of the view on-screen.

Original issue reported here:
https://issuetracker.google.com/issues/142835261

Test: added unit tests for verifying the epicenter in simple situations.

Change-Id: Iddab078201d19deac4a30a3b60d976c85fb5d7ac