Only send Behaviors ACTION_CANCEL events when necessary.
Changes CoordinatorLayout to only send ACTION_CANCEL events to views
when it is necessary based on the prior stream of events that they have
received (i.e. they have not already received a cancel event or an up
event).

Fixes a couple of existing bugs in the event dispatch;
 * behaviors prior to a behavior that is intercepting did not receive
   CANCEL events until the end of the event stream.
 * behaviors after a blocking behavior still received the event that a
   behavior blocked on.
 * behaviors were sent on touch before any had intercepted an event, now
   behaviors continue to be sent onInterceptTouchEvent during the
   CoordinatorLayout's onTouchEvent until one chooses to intercept.

Introduces a unit test to more thoroughly test the event dispatch and
cancel paths for multiple behaviors.

Bug: 143302939
Change-Id: I738a0ff1d59f3717b60b918cbb38d075d7f1c04b
Fix: 143302939
Test: ./gradlew coordinatorlayout:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.package=androidx.coordinatorlayout.widget