Ensure views are visible when popping a reordered hide
When an exiting fragment uses transitions and does a hide in a
FragmentTransaction that allows reordering, if that transaction is
popped the Fragment views are still invisible.

This is because when reordering is allowed on an exiting fragment, the
exiting views are set to INVISIBLE. When the transaction is popped,
the Fragment view instance is new, but if the child views of the
Fragment view has ids, the same views are restored. These views are
still set to INVISIBLE, and therefore these views are ignored.

This change ensures that the entering Fragment view and its children are
set to VISIBlE so they can properly do any transitions.

Test: All tests pass
BUG: 70793925
Change-Id: I4f26981afdb49ec9f262b8e46fc63147858551a5