Create lint check for null safe LiveData
LiveData values are always nullable even when explicitly declaring the
type as non-nullable in kotlin. This can result in runtime exceptions
from attempting to read a null value that was not expected to be
nullable.

This lint check enforces instantiating MutableLiveData objects with
nullable type parameters in Kotlin. Using nullable type parameters leads
to studio warnings being displayed that enforce using null safe checks
when reading the LiveData value.

Test: added NonNullableMutableLiveDataDetectorTest
Bug: 130261288
Change-Id: Ie736a98d3235a51a5400db199430b8fdd1da004e