Call cancel callbacks when removed from tree.
Actually, the cancel callback currently does not
happen when the PointerInputNode is removed from
the tree, but rather when the next pointer input
event is processed and it is discovered that the
PointerInputNode is no longer part of the tree.

This should likely not be the case and will be
addressed along side larger changes that are likely
coming to pointer input (such as turning pointer input
into modifiers).

Bug: 142486858
Test: ./gradlew ui:ui-platform:testDebugUnitTest --tests=androidx.ui.core.pointerinput.HitPathTrackerTest
TestL ./gradlew ui:ui-platform:testDebugUnitTest --tests=androidx.ui.core.pointerinput.PointerInputProcessorTest

Change-Id: I7112a02a49d54ab2176f6f916696b00d5d965912