Save NavHostFragment graphId across config change
When using FragmentContainerView to inflate a NavHostFragment if the
activity is recreated (i.e. a config change), the graphId for the
NavHostFragment is no longer available. This causes an
IllegalStateException when attempting to navigate after recreation.

Previously with the fragment tag, onInflate() would be called on every
recreation and therefore the graph would set again.
FragmentContainerView only calls onInflate() the first time. This means
that the NavHostFragment should save and restore its graphId
appropriately using onSaveInstanceState().

Test: Added in NavHostFragmentTest
BUG: 143752103
Change-Id: I90a6f1d88b77a0a78f614d158812080936deba4d